@@ -1,16 +1,18 @@
+#!/usr/bin/env python3
 """A CLI for the evohome-sync2 library."""
 
 import sys
 
 try:
-    from evohomeasync2.client import main
+    from cli.client import main
 
 except ModuleNotFoundError:
-    import os
+    from pathlib import Path
 
-    sys.path.append(f"{os.path.dirname(__file__)}/src")
+    sys.path.append(str(Path(__file__).parent / "cli"))
+    sys.path.append(str(Path(__file__).parent / "src"))
 
-    from evohomeasync2.client import main
+    from cli.client import main
 
 if __name__ == "__main__":
     main()
