@@ -1,433 +1,276 @@
-#!/usr/bin/env python3
-"""evohomeasync2 provides an async client for the updated Evohome TCC API."""
+"""evohomeasync provides an async client for the v2 Resideo TCC API."""
 
 from __future__ import annotations
 
-import logging
+import base64
 from abc import ABC, abstractmethod
-from collections.abc import Generator
-from datetime import datetime as dt, timedelta as td
-from http import HTTPMethod, HTTPStatus
-from types import TracebackType
-from typing import TYPE_CHECKING, Any, Final, TypedDict
+from datetime import UTC, datetime as dt, timedelta as td
+from http import HTTPStatus
+from typing import TYPE_CHECKING, Any, Final
 
-import aiohttp
 import voluptuous as vol
 
+from evohome.auth import AbstractAuth
+from evohome.const import HEADERS_BASE, HEADERS_CRED, HINT_BAD_CREDS
+from evohome.credentials import CredentialsManagerBase
+from evohome.helpers import convert_keys_to_snake_case, obfuscate
+
 from . import exceptions as exc
-from .const import (
-    AUTH_HEADER,
-    AUTH_HEADER_ACCEPT,
-    AUTH_PAYLOAD,
-    AUTH_URL,
-    CREDS_REFRESH_TOKEN,
-    CREDS_USER_PASSWORD,
-    SCH_OAUTH_TOKEN,
-    SZ_ACCESS_TOKEN,
-    SZ_ACCESS_TOKEN_EXPIRES,
-    SZ_EXPIRES_IN,
-    SZ_PASSWORD,
-    SZ_REFRESH_TOKEN,
-    SZ_USERNAME,
-    URL_BASE,
-    URL_HOST,
-)
 
 if TYPE_CHECKING:
+    import logging
+
+    import aiohttp
     from aiohttp.typedefs import StrOrURL
 
-    from .schema import _EvoDictT, _EvoSchemaT  # pragma: no cover
+    from .schemas.typedefs import (
+        EvoAuthTokensDictT as AccessTokenEntryT,
+        TccAuthTokensResponseT as AuthTokenResponseT,  # TCC is snake_case anyway
+    )
 
 
-_LOGGER: Final = logging.getLogger(__name__)
+_APPLICATION_ID: Final = base64.b64encode(
+    b"4a231089-d2b6-41bd-a5eb-16a0a422b999:"  # fmt: off
+    b"1a15cdb8-42de-407b-add0-059f92c530cb"
+).decode("utf-8")
 
 
-_ERR_MSG_LOOKUP_BOTH: dict[int, str] = {  # common to both OAUTH_URL & URL_BASE
-    HTTPStatus.INTERNAL_SERVER_ERROR: "Can't reach server (check vendor's status page)",
-    HTTPStatus.METHOD_NOT_ALLOWED: "Method not allowed (dev/test?)",
-    HTTPStatus.SERVICE_UNAVAILABLE: "Can't reach server (check vendor's status page)",
-    HTTPStatus.TOO_MANY_REQUESTS: "Vendor's API rate limit exceeded (wait a while)",
+SZ_ACCESS_TOKEN: Final = "access_token"  # noqa: S105
+SZ_ACCESS_TOKEN_EXPIRES: Final = "access_token_expires"  # noqa: S105
+SZ_EXPIRES_IN: Final = "expires_in"
+SZ_REFRESH_TOKEN: Final = "refresh_token"  # noqa: S105
+
+
+SCH_OAUTH_TOKEN: Final = vol.Schema(
+    {
+        vol.Required(SZ_ACCESS_TOKEN): vol.All(str, obfuscate),
+        vol.Required(SZ_EXPIRES_IN): int,  # 1800 seconds
+        vol.Required(SZ_REFRESH_TOKEN): vol.All(str, obfuscate),
+        vol.Required("token_type"): str,
+        vol.Optional("scope"): str,  # "EMEA-V1-Basic EMEA-V1-Anonymous"
+    }
+)
+
+CREDS_REFRESH_TOKEN: Final = {
+    "grant_type": "refresh_token",
+    "scope": "EMEA-V1-Basic EMEA-V1-Anonymous",
+    "refresh_token": "",
 }
 
-_ERR_MSG_LOOKUP_AUTH: dict[int, str] = _ERR_MSG_LOOKUP_BOTH | {  # POST OAUTH_URL
-    HTTPStatus.BAD_REQUEST: "Invalid user credentials (check the username/password)",
-    HTTPStatus.NOT_FOUND: "Not Found (invalid URL?)",
-    HTTPStatus.UNAUTHORIZED: "Invalid access token (dev/test only)",
-}
-
-_ERR_MSG_LOOKUP_BASE: dict[int, str] = _ERR_MSG_LOOKUP_BOTH | {  # GET/PUT URL_BASE
-    HTTPStatus.BAD_REQUEST: "Bad request (invalid data/json?)",
-    HTTPStatus.NOT_FOUND: "Not Found (invalid entity type?)",
-    HTTPStatus.UNAUTHORIZED: "Unauthorized (expired access token/unknown entity id?)",
+CREDS_USER_PASSWORD: Final = {
+    "grant_type": "password",
+    "scope": "EMEA-V1-Basic EMEA-V1-Anonymous",  # EMEA-V1-Get-Current-User-Account",
+    "Username": "",
+    "Password": "",
 }
 
 
-class OAuthTokenData(TypedDict):
-    access_token: str
-    expires_in: int  # number of seconds
-    refresh_token: str
+# POST authentication url (i.e. /Auth/OAuth/Token)
+URL_CRED: Final = "Auth/OAuth/Token"
+
+# GET/PUT resource url (e.g. /WebAPI/emea/api/v1/...)
+URL_BASE: Final = "WebAPI/emea/api/v1"
 
 
-class _EvoTokenData(TypedDict):
-    access_token: str
-    access_token_expires: str  # dt.isoformat()
-    refresh_token: str
-
-
-class AbstractTokenManager(ABC):
-    """Abstract class to manage an OAuth access token and its refresh token."""
+class AbstractTokenManager(CredentialsManagerBase, ABC):
+    """An ABC for managing the auth tokens used for HTTP authentication."""
 
     _access_token: str
     _access_token_expires: dt
-    _refresh_token: str
-    websession: aiohttp.ClientSession
+    _refresh_token: str = ""
 
     def __init__(
         self,
-        username: str,
-        password: str,
+        client_id: str,
+        secret: str,
         websession: aiohttp.ClientSession,
+        /,
+        _hostname: str | None = None,
+        logger: logging.Logger | None = None,
     ) -> None:
         """Initialize the token manager."""
 
-        self._user_credentials = {
-            SZ_USERNAME: username,
-            SZ_PASSWORD: password,
-        }
+        super().__init__(
+            client_id, secret, websession, _hostname=_hostname, logger=logger
+        )
+        self._clear_access_token()  # initialise the attrs
 
-        self.websession = websession
+    def _clear_access_token(self) -> None:
+        """Clear the auth tokens attrs (set to falsey state)."""
 
-        self._auth_tokens_clear()
-
-    def __str__(self) -> str:
-        return f"{self.__class__.__name__}(username='{self.username}')"
+        self._access_token = ""
+        self._access_token_expires = dt.min.replace(tzinfo=UTC)  # don't need local TZ
 
     @property
     def access_token(self) -> str:
-        """Return the access_token."""
+        """Return the access token."""
         return self._access_token
 
     @property
     def access_token_expires(self) -> dt:
-        """Return the access_token_expires."""
+        """Return the expiration datetime of the access token."""
         return self._access_token_expires
 
     @property
     def refresh_token(self) -> str:
-        """Return the refresh_token."""
+        """Return the refresh token."""
         return self._refresh_token
 
-    @property
-    def username(self) -> str:
-        """Return the username."""
-        return self._user_credentials[SZ_USERNAME]
+    async def get_access_token(self) -> str:  # convenience wrapper
+        """Return a valid access token.
 
-    def _auth_tokens_clear(self) -> None:
-        """Reset the token data to its falsy state."""
-        self._access_token = ""
-        self._access_token_expires = dt.min
-        self._refresh_token = ""
+        If required, fetch (and save) a new token via the vendor's web API.
+        """
 
-    def _auth_tokens_from_api(self, tokens: OAuthTokenData) -> tuple[str, dt, str]:
-        """Convert the token data from the vendor's API to the internal format."""
+        if not self.is_token_valid():  # although may be rejected for other reasons
+            await self.fetch_access_token()
+            await self.save_access_token()
 
-        self._access_token = tokens[SZ_ACCESS_TOKEN]
-        self._access_token_expires = dt.now() + td(seconds=tokens[SZ_EXPIRES_IN])
-        self._refresh_token = tokens[SZ_REFRESH_TOKEN]
+        return self.access_token
 
-        return (
-            self._access_token,
-            self._access_token_expires,
-            self._refresh_token,
+    def is_token_valid(self) -> bool:
+        """Return True if the access token is valid (the server may still reject it)."""
+        return self._access_token_expires > dt.now(tz=UTC) + td(seconds=15)
+
+    async def fetch_access_token(self) -> None:
+        """Update the access token and save it to the store/cache."""
+
+        self.logger.debug("Fetching access_token...")
+
+        if self._refresh_token:
+            self.logger.debug(" - authenticating with the refresh_token")
+
+            credentials = {SZ_REFRESH_TOKEN: self._refresh_token}
+
+            try:
+                await self._fetch_access_token(CREDS_REFRESH_TOKEN | credentials)
+
+            except exc.AuthenticationFailedError as err:
+                if err.status != HTTPStatus.BAD_REQUEST:  # 400, e.g. invalid tokens
+                    raise
+
+                self.logger.debug("Expired/invalid refresh_token")
+                self._refresh_token = ""
+
+        if not self._refresh_token:
+            self.logger.debug(" - authenticating with client_id/secret")
+
+            credentials = {
+                # NOTE: the keys are case-sensitive: 'Username' and 'Password'
+                "Username": self._client_id,
+                "Password": self._secret,
+            }
+
+            try:  # check here if the user credentials are invalid...
+                await self._fetch_access_token(CREDS_USER_PASSWORD | credentials)
+
+            except exc.AuthenticationFailedError as err:
+                if "invalid_grant" not in err.message:
+                    raise
+                self.logger.error(HINT_BAD_CREDS)  # noqa: TRY400
+                # status will be: HTTPStatus.BAD_REQUEST, 400
+                raise exc.BadUserCredentialsError(f"{err}", status=err.status) from err
+
+            self._was_authenticated = True
+
+        self.logger.debug(f" - access_token = {self.access_token}")
+        self.logger.debug(f" - access_token_expires = {self.access_token_expires}")
+        self.logger.debug(f" - refresh_token = {self.refresh_token}")
+
+    async def _fetch_access_token(self, credentials: dict[str, str]) -> None:
+        """Obtain an access token using the supplied credentials.
+
+        The credentials are either a refresh token or the user's client_id/secret.
+        Will raise AuthenticationFailedError if unable to obtain an access token.
+        """
+
+        url = f"https://{self.hostname}/{URL_CRED}"
+
+        response: AuthTokenResponseT = await self._post_access_token_request(
+            url,
+            headers=HEADERS_CRED | {"Authorization": "Basic " + _APPLICATION_ID},
+            data=credentials,  # NOTE: is snake_case
         )
 
-    def _deserialize_auth_tokens(self, tokens: _EvoTokenData) -> None:
-        """Deserialize the token data from a dictionary."""
+        try:  # the dict _should_ be the expected schema...
+            self.logger.debug(
+                f"POST {url}: {SCH_OAUTH_TOKEN(response)}"  # should be obfuscated
+            )
+
+        except vol.Invalid as err:
+            self.logger.warning(f"POST {url}: payload may be invalid: {err}")
+
+        tokens: AuthTokenResponseT = convert_keys_to_snake_case(response)
+
+        try:
+            self._access_token = tokens[SZ_ACCESS_TOKEN]
+            self._access_token_expires = dt.now(tz=UTC) + td(
+                seconds=tokens[SZ_EXPIRES_IN]
+            )
+            self._refresh_token = tokens[SZ_REFRESH_TOKEN]
+
+        except (KeyError, TypeError) as err:
+            raise exc.AuthenticationFailedError(
+                f"Authenticator response is invalid: {err}, payload={response}"
+            ) from err
+
+        self._was_authenticated = True  # i.e. the credentials are valid
+
+        # if response.get(SZ_LATEST_EULA_ACCEPTED):
+
+    async def _post_access_token_request(  # dev/test wrapper (also typing)
+        self, url: StrOrURL, /, **kwargs: Any
+    ) -> AuthTokenResponseT:
+        """Wrap the POST request to the vendor's TCC RESTful API."""
+        return await self._post_request(url, **kwargs)  # type: ignore[return-value]
+
+    @abstractmethod
+    async def save_access_token(self) -> None:
+        """Save the (serialized) access token (and expiry dtm, refresh token).
+
+        Should ideally confirm the access token is valid before saving.
+        """
+
+    def _import_access_token(self, tokens: AccessTokenEntryT) -> None:
+        """Extract the token data from a (serialized) dictionary."""
+
         self._access_token = tokens[SZ_ACCESS_TOKEN]
         self._access_token_expires = dt.fromisoformat(tokens[SZ_ACCESS_TOKEN_EXPIRES])
         self._refresh_token = tokens[SZ_REFRESH_TOKEN]
 
-    def _serialize_auth_tokens(self) -> _EvoTokenData:
-        """Serialize the token data to a dictionary."""
+    def _export_access_token(self) -> AccessTokenEntryT:
+        """Convert the token data to a (serialized) dictionary."""
+
         return {
-            SZ_ACCESS_TOKEN: self.access_token,
-            SZ_ACCESS_TOKEN_EXPIRES: self.access_token_expires.isoformat(),
-            SZ_REFRESH_TOKEN: self.refresh_token,
+            SZ_ACCESS_TOKEN_EXPIRES: self._access_token_expires.isoformat(),
+            SZ_ACCESS_TOKEN: self._access_token,
+            SZ_REFRESH_TOKEN: self._refresh_token,
         }
-
-    def is_token_data_valid(self) -> bool:
-        """Return True if we have a valid access token."""
-        return bool(self.access_token) and self.access_token_expires > dt.now()
-
-    async def get_access_token(self) -> str:
-        """Return a valid access token.
-
-        If required, fetch a new token via the vendor's web API.
-        """
-
-        if not self.is_token_data_valid():  # TODO: but may be invalid for other reasons
-            _LOGGER.warning("Missing/Expired/Invalid access_token, re-authenticating.")
-            await self._update_access_token()
-
-        return self.access_token
-
-    async def _update_access_token(self) -> None:
-        """Update the access token and save it to the store/cache."""
-
-        if self._refresh_token:
-            _LOGGER.warning("Authenticating with the refresh_token...")
-
-            try:
-                await self._obtain_access_token(
-                    CREDS_REFRESH_TOKEN | {SZ_REFRESH_TOKEN: self.refresh_token}
-                )
-
-            except exc.AuthenticationFailedError as err:
-                if err.status != HTTPStatus.BAD_REQUEST:  # e.g. invalid tokens
-                    raise
-
-                _LOGGER.warning(" - invalid/expired refresh_token")
-                self._refresh_token = ""
-
-        if not self._refresh_token:
-            _LOGGER.warning("Authenticating with username/password...")
-
-            await self._obtain_access_token(
-                CREDS_USER_PASSWORD | self._user_credentials
-            )
-
-        await self.save_access_token()
-
-        _LOGGER.debug(f" - refresh_token = {self.refresh_token}")
-        _LOGGER.debug(f" - access_token = {self.access_token}")
-        _LOGGER.debug(f" - access_token_expires = {self.access_token_expires}")
-
-    async def _obtain_access_token(self, credentials: dict[str, str]) -> None:
-        """Obtain an access token using the supplied credentials.
-
-        The credentials are either a refresh token or the client_id/secret.
-        """
-
-        token_data = await self._post_access_token_request(
-            AUTH_URL,
-            data=AUTH_PAYLOAD | credentials,
-            headers=AUTH_HEADER,
-        )
-
-        try:  # the access token _should_ be valid...
-            _ = SCH_OAUTH_TOKEN(token_data)  # can't use this result, due to obsfucation
-        except vol.Invalid as err:
-            _LOGGER.warning(
-                f"Response JSON may be invalid: POST {AUTH_URL}: vol.Invalid({err})"
-            )
-
-        try:
-            self._auth_tokens_from_api(token_data)
-        except (KeyError, TypeError) as err:
-            raise exc.AuthenticationFailedError(
-                f"Invalid response from server: {err}"
-            ) from err
-
-    async def _post_access_token_request(
-        self, url: str, **kwargs: Any
-    ) -> OAuthTokenData:
-        """Obtain an access token via the vendor's web API."""
-
-        try:
-            async with self.websession.post(url, **kwargs) as response:
-                response.raise_for_status()
-
-                return await response.json()  # type: ignore[no-any-return]
-
-        except aiohttp.ContentTypeError as err:
-            # <title>Authorize error <h1>Authorization failed
-            # <p>The authorization server have encoutered an error while processing...  # codespell:ignore encoutered
-            content = await response.text()
-            raise exc.AuthenticationFailedError(
-                f"Server response is not JSON: {HTTPMethod.POST} {AUTH_URL}: {content}"
-            ) from err
-
-        except aiohttp.ClientResponseError as err:
-            if hint := _ERR_MSG_LOOKUP_AUTH.get(err.status):
-                raise exc.AuthenticationFailedError(hint, status=err.status) from err
-            raise exc.AuthenticationFailedError(str(err), status=err.status) from err
-
-        except aiohttp.ClientError as err:  # e.g. ClientConnectionError
-            raise exc.AuthenticationFailedError(str(err)) from err
-
-    @abstractmethod
-    async def save_access_token(self) -> None:  # HA: api
-        """Save the access token to a cache."""
-
-
-class _RequestContextManager:
-    """A context manager for Auth's aiohttp request."""
-
-    _response: aiohttp.ClientResponse | None = None
-
-    def __init__(
-        self,
-        websession: aiohttp.ClientSession,
-        method: HTTPMethod,
-        url: StrOrURL,
-        /,
-        **kwargs: Any,
-    ):
-        """Initialize the request context manager."""
-
-        self.websession = websession
-        self.method = method
-        self.url = url
-        self.kwargs = kwargs
-
-    async def __aenter__(self) -> aiohttp.ClientResponse:
-        """Async context manager entry."""
-        self._response = await self._await_impl()
-        return self._response
-
-    async def __aexit__(
-        self,
-        exc_type: type[BaseException] | None,
-        exc: BaseException | None,
-        tb: TracebackType | None,
-    ) -> None:
-        """Async context manager exit."""
-        if self._response:
-            self._response.release()
-            await self._response.wait_for_close()
-
-    def __await__(self) -> Generator[Any, Any, aiohttp.ClientResponse]:
-        """Make this class awaitable."""
-        return self._await_impl().__await__()
-
-    async def _await_impl(self) -> aiohttp.ClientResponse:
-        """Return the actual result."""
-        return await self.websession.request(self.method, self.url, **self.kwargs)
-
-
-class AbstractAuth(ABC):  # APIs esposed by/for HA
-    def __init__(self, websession: aiohttp.ClientSession, host: str) -> None:
-        """Initialize the auth."""
-        self.websession = websession
-        self.host = host
-
-    @abstractmethod
-    async def get_access_token(self) -> str:
-        """Return a valid access token."""
-
-    async def request(  # type: ignore[no-untyped-def]
-        self, method: HTTPMethod, url: StrOrURL, /, **kwargs: Any
-    ):
-        """Make a request to the Evohome TCC API."""
-
-        headers = kwargs.pop("headers", None) or {
-            "Accept": AUTH_HEADER_ACCEPT,
-            "Content-Type": "application/json",
-        }
-        headers["Authorization"] = "bearer " + await self.get_access_token()
-
-        return await _RequestContextManager(
-            self.websession, method, url, **kwargs, headers=headers
-        )
 
 
 class Auth(AbstractAuth):
-    """A class for interacting with the Evohome TCC API."""
+    """A class for interacting with the v2 Resideo TCC API."""
 
     def __init__(
         self,
-        websession: aiohttp.ClientSession,
         token_manager: AbstractTokenManager,
-        logger: logging.Logger,
+        websession: aiohttp.ClientSession,
+        /,
+        *,
+        _hostname: str | None = None,
+        logger: logging.Logger | None = None,
     ) -> None:
-        """A class for interacting with the v2 Evohome TCC API."""
+        """A class for interacting with the v2 Resideo TCC API."""
 
-        super().__init__(websession, URL_HOST)
+        super().__init__(websession, _hostname=_hostname, logger=logger)
 
-        self.token_manager = token_manager
-        self._logger = logger
+        self._access_token = token_manager.get_access_token
+        self._url_base = f"https://{self.hostname}/{URL_BASE}"
 
-    async def get_access_token(self) -> str:
-        """Return a valid access token."""
-        return await self.token_manager.get_access_token()
+    async def _headers(self, headers: dict[str, str] | None = None) -> dict[str, str]:
+        """Ensure the authorization header has a valid access token."""
 
-    async def _request(
-        self, method: HTTPMethod, url: StrOrURL, /, **kwargs: Any
-    ) -> dict[str, Any] | list[dict[str, Any]] | str | None:
-        """Make a request to the Evohome TCC API."""
-
-        async with await self.request(method, url, **kwargs) as response:
-            try:
-                response.raise_for_status()
-
-            except aiohttp.ClientResponseError as err:
-                if hint := _ERR_MSG_LOOKUP_BASE.get(err.status):
-                    raise exc.RequestFailedError(hint, status=err.status) from err
-                raise exc.RequestFailedError(str(err), status=err.status) from err
-
-            except aiohttp.ClientError as err:  # e.g. ClientConnectionError
-                raise exc.RequestFailedError(str(err)) from err
-
-            return await self._content(response)
-
-    @staticmethod
-    async def _content(
-        response: aiohttp.ClientResponse,
-    ) -> dict[str, Any] | list[dict[str, Any]] | str | None:
-        """Return the content of the response."""
-
-        if not response.content_length:
-            return None
-
-        if response.content_type != "application/json":
-            # assume "text/plain" or "text/html"
-            return await response.text()
-
-        content: dict[str, Any] = await response.json()
-        return content
-
-    async def get(self, url: StrOrURL, schema: vol.Schema | None = None) -> _EvoSchemaT:
-        """Call the Evohome TCC API with a GET.
-
-        Optionally checks the response JSON against the expected schema and logs a
-        warning if it doesn't match (NB: does not raise a vol.Invalid).
-        """
-
-        content: _EvoSchemaT
-
-        content = await self._request(  # type: ignore[assignment]
-            HTTPMethod.GET, f"{URL_BASE}/{url}"
-        )
-
-        if schema:
-            try:
-                content = schema(content)
-            except vol.Invalid as err:
-                self._logger.warning(
-                    f"Response JSON may be invalid: GET {url}: vol.Invalid({err})"
-                )
-
-        return content
-
-    async def put(
-        self, url: StrOrURL, json: _EvoDictT | str, schema: vol.Schema | None = None
-    ) -> dict[str, Any] | list[dict[str, Any]]:  # NOTE: not _EvoSchemaT
-        """Call the Evohome TCC API with a PUT (POST is only used for authentication).
-
-        Optionally checks the request JSON against the expected schema and logs a
-        warning if it doesn't match (NB: does not raise a vol.Invalid).
-        """
-
-        content: dict[str, Any] | list[dict[str, Any]]
-
-        if schema:
-            try:
-                _ = schema(json)
-            except vol.Invalid as err:
-                self._logger.warning(f"Request JSON may be invalid: PUT {url}: {err}")
-
-        content = await self._request(  # type: ignore[assignment]
-            HTTPMethod.PUT, f"{URL_BASE}/{url}", json=json
-        )
-
-        return content
+        headers = HEADERS_BASE | (headers or {})
+        return headers | {
+            "Authorization": "bearer " + await self._access_token(),
+        }
