@@ -1,31 +1,30 @@
-#!/usr/bin/env python3
 """Tests for evohome-async - validate the schema of HA's debug JSON (newer ver)."""
 
 from __future__ import annotations
 
-import json
 from pathlib import Path
+from typing import TYPE_CHECKING, Any, Final
 
-import pytest
-
-from evohomeasync2 import Location
-from evohomeasync2.schema import SCH_LOCN_STATUS
-from evohomeasync2.schema.config import SCH_TCS_CONFIG, SCH_TIME_ZONE
-from evohomeasync2.schema.const import (
+from evohome.helpers import convert_keys_to_snake_case
+from evohomeasync2.location import create_location
+from evohomeasync2.schemas import TCC_GET_LOC_STATUS
+from evohomeasync2.schemas.config import factory_tcs, factory_time_zone
+from evohomeasync2.schemas.const import (
     S2_GATEWAYS,
     S2_LOCATION_INFO,
+    S2_SUPPORTS_DAYLIGHT_SAVING,
     S2_TEMPERATURE_CONTROL_SYSTEMS,
     S2_TIME_ZONE,
+    S2_USE_DAYLIGHT_SAVE_SWITCHING,
 )
 
 from .conftest import ClientStub
-from .helpers import TEST_DIR
+from .const import TEST_DIR
+
+if TYPE_CHECKING:
+    import pytest
 
 WORK_DIR = f"{TEST_DIR}/schemas_1"
-
-# NOTE: JSON from HA is not compliant with vendor schema, but is useful to test against
-CONFIG_FILE_NAME = "config.json"
-STATUS_FILE_NAME = "status.json"
 
 
 def pytest_generate_tests(metafunc: pytest.Metafunc) -> None:
@@ -38,34 +37,30 @@ def pytest_generate_tests(metafunc: pytest.Metafunc) -> None:
     metafunc.parametrize("folder", sorted(folders), ids=id_fnc)
 
 
-def test_config_refresh(folder: Path) -> None:
+# These schemas have camelCase keys, as per the vendor's schema
+SCH_TCS_CONFIG: Final = factory_tcs()
+SCH_TIME_ZONE: Final = factory_time_zone()
+
+
+async def test_config_refresh(config: dict[str, Any], status: dict[str, Any]) -> None:
     """Test the loading a config, then an update_status() on top of that."""
 
-    if not Path(folder).joinpath(CONFIG_FILE_NAME).is_file():
-        pytest.skip(f"No {CONFIG_FILE_NAME} in: {folder.name}")
+    # hack because old JSON from HA's evohome integration didn't include this data
+    if config[S2_LOCATION_INFO].get(S2_USE_DAYLIGHT_SAVE_SWITCHING) is None:
+        config[S2_LOCATION_INFO][S2_USE_DAYLIGHT_SAVE_SWITCHING] = config[
+            S2_LOCATION_INFO
+        ][S2_TIME_ZONE][S2_SUPPORTS_DAYLIGHT_SAVING]
 
-    if not Path(folder).joinpath(STATUS_FILE_NAME).is_file():
-        pytest.skip(f"No {STATUS_FILE_NAME} in: {folder.name}")
-
-    with open(Path(folder).joinpath(CONFIG_FILE_NAME)) as f:
-        config: dict = json.load(f)  # is camelCase, as per vendor's schema
-
-    with open(Path(folder).joinpath(STATUS_FILE_NAME)) as f:
-        status: dict = json.load(f)  # is camelCase, as per vendor's schema
+    config = convert_keys_to_snake_case(config)
+    status = convert_keys_to_snake_case(status)
 
     # for this, we need snake_case keys
-    loc = Location(ClientStub(), config)
-    loc._update_status(status)
+    loc = await create_location(ClientStub(), config)  # type: ignore[arg-type]
+    loc._update_status(status)  # type: ignore[arg-type]
 
 
-def test_config_schemas(folder: Path) -> None:
+def test_config_schemas(config: dict[str, Any]) -> None:
     """Test the config schema for a location."""
-
-    if not Path(folder).joinpath(CONFIG_FILE_NAME).is_file():
-        pytest.skip(f"No {CONFIG_FILE_NAME} in: {folder.name}")
-
-    with open(Path(folder).joinpath(CONFIG_FILE_NAME)) as f:
-        config: dict = json.load(f)  # is camelCase, as per vendor's schema
 
     _ = SCH_TIME_ZONE(config[S2_LOCATION_INFO][S2_TIME_ZONE])
 
@@ -74,13 +69,7 @@ def test_config_schemas(folder: Path) -> None:
             _ = SCH_TCS_CONFIG(tcs_config)
 
 
-def test_status_schemas(folder: Path) -> None:
+def test_status_schemas(status: dict[str, Any]) -> None:
     """Test the status schema for a location."""
 
-    if not Path(folder).joinpath(STATUS_FILE_NAME).is_file():
-        pytest.skip(f"No {STATUS_FILE_NAME} in: {folder.name}")
-
-    with open(Path(folder).joinpath(STATUS_FILE_NAME)) as f:
-        status: dict = json.load(f)  # is camelCase, as per vendor's schema
-
-    _ = SCH_LOCN_STATUS(status)
+    _ = TCC_GET_LOC_STATUS(status)
