@@ -1,17 +1,23 @@
-#!/usr/bin/env python3
 """Tests for evohome-async - validate the schemas of vendor's RESTful JSON."""
 
 from __future__ import annotations
 
 from pathlib import Path
+from typing import TYPE_CHECKING
 
-import pytest
+from evohomeasync2.schemas import (
+    TCC_GET_USR_ACCOUNT,
+    TCC_GET_USR_LOCATIONS,
+    factory_loc_status,
+)
 
-from evohomeasync2.schema import SCH_FULL_CONFIG, SCH_LOCN_STATUS, SCH_USER_ACCOUNT
+from .common import assert_schema
+from .const import TEST_DIR
 
-from .helpers import TEST_DIR, test_schema
+if TYPE_CHECKING:
+    import pytest
 
-WORK_DIR = f"{TEST_DIR}/systems_0"
+WORK_DIR = f"{TEST_DIR}/fixtures"
 
 
 def pytest_generate_tests(metafunc: pytest.Metafunc) -> None:
@@ -26,15 +32,20 @@ def pytest_generate_tests(metafunc: pytest.Metafunc) -> None:
 
 def test_user_account(folder: Path) -> None:
     """Test the user account schema against the corresponding JSON."""
-    test_schema(folder, SCH_USER_ACCOUNT, "user_account.json")
+
+    assert_schema(folder, TCC_GET_USR_ACCOUNT, "user_account.json")
 
 
 def test_user_locations(folder: Path) -> None:
     """Test the user locations config schema against the corresponding JSON."""
-    test_schema(folder, SCH_FULL_CONFIG, "user_locations.json")
+
+    assert_schema(folder, TCC_GET_USR_LOCATIONS, "user_locations.json")
 
 
 def test_location_status(folder: Path) -> None:
     """Test the location status schema against the corresponding JSON."""
+
+    SCH_STATUS = factory_loc_status()
+
     for p in Path(folder).glob("status_*.json"):
-        test_schema(folder, SCH_LOCN_STATUS, p.name)
+        assert_schema(folder, SCH_STATUS, p.name)
